+++ RogueLike/scripts/RenderBatchComponent.cpp
@@ -82,7 +82,7 @@ void RenderBatchComponent::initialize(const std::shared_ptr<ModelData> _modelDat
 	this->m_Material = _materialToUse;
 	this->m_ModelData = _modelDataToUse;
 	this->m_MatrixBuffer = new glm::mat4[_objectsToBatch.size()];
-	for (int i = 0; i < _objectsToBatch.size(); i++) {
+	for (int i = 0; i < (int) _objectsToBatch.size(); i++) {
 		_objectsToBatch[i]->getTransform().setDirty();
 		_objectsToBatch[i]->getTransform().updateCurrentMatrix();
 		m_MatrixBuffer[i] = _objectsToBatch[i]->getTransform().getObjectMatrix();
